Revert "ARM: 8430/1: use default ioremap alignment for SMP or LPAE"
This reverts commit 803e3dbcb4cf80c898faccf01875f6ff6e5e76fd.

With this change the zynq pl353 nand controller is not working
because the ioremap function returning the address that is  not
aligned to the given size.

Still debugging why the nand transactions are failing even if the
virtual address is not aligned to given size.

So, reverting the change till the resolution is found for the above
issue.

Signed-off-by: Michal Simek <michal.simek@xilinx.com>