mtd: nand: fix for mtd nand_oob test
mtd nand oob test gives reset to nand device at the start.
This will be problem for the nand devices which are configured for
nvddr modes. i.e controller and device both are not is same mode,
resulting failure in mtd-oob test. this patch fixes this with out
issuing reset.

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>