mtd: spi-nor: reset EAR to 0 on reboot/shutdown
The boot ROM on the Zynq does not 'know' the EAR register, so
if the EAR is set to non-zero and the system reboots, the ROM
is unable to find the boot sector and the system hangs.

This adds a shutdown handler to the spi-nor driver to set the
EAR register back to 0 at shutdown or reboot time.

Signed-off-by: Ranjit Waghmode <ranjit.waghmode@xilinx.com>
Reviewed-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>