Revert "Revert "ARM: 8430/1: use default ioremap alignment for SMP or LPAE""
This reverts commit "68e7a641fba1adeacdb766aa8620ed6f69a142f9".

There is a kernel patch which fixes the use of default ioremap alignment
for SMP or LPAE ("803e3dbcb4cf80c898faccf01875f6ff6e5e76fd"). With this
patch our pl35x_nand driver got broken because the ioremap function
returning the address that is not aligned to the given size. hence we
reverted this patch till we fix the driver.

Now the driver got updated to resolve the above issue, hence this patch
reverts to original commit "803e3dbcb4cf".

Signed-off-by: Naga Sureshkumar Relli <nagasure@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>