phy: zynqmp: reinitialize gt lane settings for usb only if required
As a part of initializing the GT lanes for USB, the USB controller
should be kept into reset and then perform the initialization.
But doing so will make the USB lose the saved context which was
stored before suspend when hibernation feature is enabled in USB.
In order to avoid the above issue, the driver should avoid calling
phy_init() for every suspend & resume sequence, except when FPD
power rails are cut. As per ZynqMP architecture, usb hibernation is
not intended to work with FPD off scenario. The phy_init() should be
called for USB after resume only if FPD power rails were cut during
suspend or when the peripheral driver changes the GT lane from the
previous configured one.

This patch uses ICM CFG registers to determine whether phy_init() can
be called or not after resume. As ICM Configuration represents protocol
that got configured for the GT lane, which remains same until explicitly
requested to change by the peripheral driver or reset to 0 when the FPD
power rails are cut during suspend (As SERDES is in the FPD domain, cutting
FPD power will remove power for SERDES phy and resets the value to default
zero). So, this patch modifies the driver to save the old ICM CFG registers
during suspend and compare it with new ICM CFG register values after
getting resumed. The GT lane initialization is performed for USB when the
saved ICM CFG reg values and the current ICM CFG reg values are different,
else gt lane initialization is skipped.

Signed-off-by: Anurag Kumar Vulisha <anuragku@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>