v4l: xilinx: dma: Fix v4l2 enumeration callback for multiplanar
This fixes issues with v4l2 enumeration callback with multiplanar
formats due to which supported multiplanar formats were not getting
enumerated when using VIDEO_ENUM_FMT ioctl.

The intention of existing implementation for multiplanar formats was
to save in array and return the subset of all the v4l2 formats
supported by attached dma device and the supported media bus format
and also cache the supported media bus format so that for future calls
if same media bus format is called then the v4l2 pixel formats can be
directly returned from the entries saved in the array.

This is acheived now with below changes :
a. Use V4L2 subdev helper functions to query supported media bus
  format code.
b. Cache the media bus format also for the condition when it is empty
  along with whenever new media bus format is detected.
c. Fix NULL pointer deference error by dynamically allocating
  memory for poss_v4l2_fmts.

Signed-off-by: Devarsh Thakkar <devarsht@xilinx.com>
Signed-off-by: Satish Kumar Nagireddy <satishna@xilinx.com>
Reviewed-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>