microblaze: adi_mb_defconfig: fix firmware build
Without this change this error occurs:
```
make[1]: *** No rule to make target
	'/lib/firmware/ad9467_intbypass_ad9517.stp',
	needed by 'firmware/ad9467_intbypass_ad9517.stp.gen.o'.  Stop.
```

The kernel has cleaned up it's `firmware` sub-directory since 4.9 and all
firmwares should now be moved to the `linux-firmware` repo.
In the meantime, external firmwares are still permitted, so make point
these firmware to a "external" folder [which is the same folder in the
kernel tree].

This should not have worked in the previous versions of the kernel, as the
default CONFIG_EXTERNAL_FIRWMARE_DIR should have been `/lib/firmware` as
well.
It was probably a fluke between file dependencies.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>